feat: bottom panel stylization
