+++ src/js/components/BottomPanel/BottomPanel.scss
@@ -1,0 +1,18 @@
+.bottom-panel {
+ display: flex;
+ align-items: center;
+ justify-content: center;
+
+ &__column {
+  flex: 1;
+  padding: 5px;
+
+  &:first-child {
+   text-align: left;
+  }
+
+  &:last-child {
+   text-align: right;
+  }
+ }
+}
